* Fixed installation destination for multi-Disc based games.   Second ISO is now properly going into the existing folder instead   of using the Disc's internal Game Title. * Prevent deleting a folder when installation is canceled if   the folder still contains another disc number. * Added a new setting to select GameCube game source: SD, USB, Auto   Auto = based on DIOS MIOS (Lite).   This new setting fix two known problems:   - Device selection for Devolution is no more dependent of     the installed DM(L) type. (Select manually before launching a game)   - Installation of a game on a device if the game is already     on another device is now possible.

+++ source/GameCube/GCDumper.cpp
@@ -201,7 +201,7 @@ int GCDumper::ReadDiscInfo(const u64 &game_offset)
 	return 0;
 }
 
-s32 GCDumper::InstallGame(const char *installpath, u32 game)
+s32 GCDumper::InstallGame(const char *installpath, u32 game, const char *installedGamePath)
 {
 	if(!ReadBuffer || game >= discHeaders.size() || game >= gameOffsets.size() || game >= gameSizes.size())
 		return -1;
@@ -265,6 +265,10 @@ s32 GCDumper::InstallGame(const char *installpath, u32 game)
 	char gamepath[512];
 	// snprintf(gamepath, sizeof(gamepath), "%s%s [%.6s]%s/", installpath, gametitle, gcheader.id, Disc ? "2" : ""); // Disc2 currently needs to be on the same folder.
 	snprintf(gamepath, sizeof(gamepath), "%s%s [%.6s]/", installpath, gametitle, gcheader.id);
+
+	// If another Disc from the same gameID already exists, let's use that path
+	if(strlen((char *)installedGamePath) != 0)
+		snprintf(gamepath, sizeof(gamepath), "%s/", installedGamePath);
 
 	CreateSubfolder(gamepath);
 
@@ -384,9 +388,12 @@ s32 GCDumper::InstallGame(const char *installpath, u32 game)
 	if(result < 0)
 	{
 		RemoveFile(gamepath);
-		char *pathPtr = strrchr(gamepath, '/');
-		if(pathPtr) *pathPtr = 0;
-		RemoveFile(gamepath);
+		if(strlen((char *)installedGamePath) == 0) // If no other disc is installed in that folder, delete it.
+		{
+			char *pathPtr = strrchr(gamepath, '/');
+			if(pathPtr) *pathPtr = 0;
+			RemoveFile(gamepath);
+		}
 
 		if(result != PROGRESS_CANCELED)
 			ShowError(tr("Disc read error."));
